+++ crates/tests/src/std_lib.rs
@@ -25,3 +25,23 @@ std.get({a: 1}, "a", error "no")
   )
   .check();
 }
+
+#[test]
+fn prune() {
+  JsonnetInput::eval_error(
+    r#"
+local xs = [2, 4, null, 6, {}];
+##    ^^ type: array[null | number | {}]
+local ys = std.prune(xs);
+##    ^^ type: array[number]
+local sc1 = std.prune({});
+##    ^^^ type: {}
+local sc2 = std.prune(null);
+##    ^^^ type: null
+
+[std.length(ys), sc1, sc2]
+"#,
+    "not yet implemented: prune",
+  )
+  .check();
+}